+++ lithops/serverless/backends/k8s/k8s.py
@@ -47,6 +47,7 @@ class KubernetesBackend:
     def __init__(self, k8s_config, internal_storage):
         logger.debug("Creating Kubernetes Job client")
         self.name = 'k8s'
+        self.type = 'batch'
         self.k8s_config = k8s_config
         self.internal_storage = internal_storage
 
@@ -214,6 +215,7 @@ class KubernetesBackend:
                                                      propagation_policy='Background')
             except Exception:
                 pass
+        self.jobs = []
 
     def list_runtimes(self, docker_image_name='all'):
         """
@@ -267,11 +269,7 @@ class KubernetesBackend:
         Invoke -- return information about this invocation
         For array jobs only remote_invocator is allowed
         """
-
         idgiver_ip = self._start_id_giver(docker_image_name)
-
-        job_payload.pop('remote_invoker')
-        job_payload.pop('invokers')
 
         executor_id = job_payload['executor_id']
         job_id = job_payload['job_id']