+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - 
 
 ### Changed
+- [Core] New monitoring system
+- [Core] New invoker system
 - [Core] Deleted strong dependency to pika==0.13.1
 
 ### Fixes